Added line rendering and camera frustum lines

+++ Hazard/src/Hazard/Core/Core.h
@@ -25,7 +25,7 @@
 #endif // HZR_DEBUG
 
 
-#define BIT(x) (1 << (x))
+#define BIT(x) (uint32_t)(1 << (x))
 #define BIND_EVENT(x) std::bind(&x, this, std::placeholders::_1)
 #define CHECK_BIT(val, bit) val & (1 << bit);
